backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
numpy==2.3.2
pandas==2.3.1

# Database dependencies
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication dependencies
bcrypt==4.1.2
PyJWT==2.8.0

# Environment and configuration
python-dotenv==1.0.0

# Validation and utilities
marshmallow==3.20.2